+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file. The format 
 
 <hr />
 
+## [2.1.42] - 2025-12-25
+Summary: Devnotes.
+
+### Changed
+- Touched areas: devnotes.
+
+### Files
+- beseo-devnotes.json
+- includes/admin/page-help-text.php
+- includes/admin/page-tools.php
+- includes/admin/playfair-admin.php
+
+
 ## [2.1.41] - 2025-12-25
 Summary: Update.